Moved ListableSettings into an own component
